reorganize modules for consistency
