+++ zebra-state/src/service/finalized_state/sled_format.rs
