--- zebra-state/src/service/memory_state.rs
@@ -1,11 +1,0 @@
-//! Non-finalized chain state management as defined by [RFC0005]
-//!
-//! [RFC0005]: https://zebra.zfnd.org/dev/rfcs/0005-state-updates.html
-
-mod chain;
-mod non_finalized_state;
-mod queued_blocks;
-
-use chain::Chain;
-pub use non_finalized_state::NonFinalizedState;
-pub use queued_blocks::QueuedBlocks;
